refactor: user DTO
