+++ src/components/users/dto/createUser.dto.ts
@@ -1,0 +1,8 @@
+import { IsNotEmpty } from 'class-validator';
+
+export class CreateUserDto {
+  @IsNotEmpty()
+  login: string;
+  @IsNotEmpty()
+  password: string;
+}
