+++ src/components/users/dto/updateUser.dto.ts
@@ -1,0 +1,8 @@
+import { IsNotEmpty } from 'class-validator';
+
+export class UpdatePasswordDto {
+  @IsNotEmpty()
+  oldPassword: string;
+  @IsNotEmpty()
+  newPassword: string;
+}
