+++ src/components/users/user.interface.ts
@@ -1,5 +1,3 @@
-import { IsNotEmpty } from 'class-validator';
-
 export interface User {
   id: string; // uuid v4
   login: string;
@@ -15,18 +13,4 @@ export interface ShowingUser {
   version: number; // integer number, increments on update
   createdAt: number; // timestamp of creation
   updatedAt: number; // timestamp of last update
-}
-
-export class CreateUserDto {
-  @IsNotEmpty()
-  login: string;
-  @IsNotEmpty()
-  password: string;
-}
-
-export class UpdatePasswordDto {
-  @IsNotEmpty()
-  oldPassword: string; // previous password
-  @IsNotEmpty()
-  newPassword: string; // new password
 }
